1.6.1

+++ src/shader/circle.shader.glsl
@@ -3,7 +3,7 @@
 */
 
 @vertex@
-precision mediump float;
+precision highp float;
 
 attribute vec2 aPosition;
 
@@ -12,13 +12,13 @@ uniform mat3 uProjection;
 varying vec2 vPosition;
 
 void main() {
-  gl_Position = vec4( ( uProjection * vec3( aPosition, 1.0 ) ).xy, 0.0, 1.0 );
-  vPosition = aPosition;
+    gl_Position = vec4( ( uProjection * vec3( aPosition, 1.0 ) ).xy, 0.0, 1.0 );
+    vPosition = aPosition;
 }
 @vertex-end@
 
 @fragment@
-precision mediump float;
+precision highp float;
 
 uniform lowp vec4 uFillColor;
 uniform lowp vec4 uOutlineColor;
@@ -28,14 +28,14 @@ uniform float uOutline;
 varying vec2 vPosition;
 
 void main() {
-  float d = dot( vPosition, vPosition );
-  float rs = uRadius * uRadius;
-  float ors = uRadius - uOutline;
-  ors = ors * ors;
-  if ( d < rs && d > ors ) {
-    gl_FragColor = uOutlineColor;
-  } else if ( d < rs ) {
-    gl_FragColor = uFillColor;
-  }
+    float d = dot( vPosition, vPosition );
+    float rs = uRadius * uRadius;
+    float ors = uRadius - uOutline;
+    ors = ors * ors;
+    if ( d < rs && d > ors ) {
+      gl_FragColor = uOutlineColor;
+    } else if ( d < rs ) {
+      gl_FragColor = uFillColor;
+    }
 }
 @fragment-end@
