+++ src/shader/line2d.shader.glsl
@@ -3,7 +3,7 @@
 */
 
 @vertex@
-precision mediump float;
+precision highp float;
 
 attribute vec2 aPosition;
 
@@ -11,16 +11,16 @@ uniform mat3 uProjection;
 uniform lowp vec4 uColor;
 
 void main() {
-  gl_Position = vec4( ( uProjection * vec3( aPosition, 1.0 ) ).xy, 0.0, 1.0 );
+    gl_Position = vec4( ( uProjection * vec3( aPosition, 1.0 ) ).xy, 0.0, 1.0 );
 }
 @vertex-end@
 
 @fragment@
-precision mediump float;
+precision highp float;
 
 uniform lowp vec4 uColor;
 
 void main() {
-  gl_FragColor = uColor;
+    gl_FragColor = uColor;
 }
 @fragment-end@
