+++ src/shader/rendertexture-sepia.shader.glsl
@@ -3,7 +3,7 @@
 */
 
 @vertex@
-precision mediump float;
+precision highp float;
 
 attribute vec2 aPosition;
 attribute vec2 aTexCoord;
@@ -23,18 +23,18 @@ void main() {
 #define TEX texture
 #endif
 
-precision mediump float;
+precision highp float;
 
 uniform sampler2D uImage;
 
 varying vec2 vTexCoord;
 
 void main() {
-  vec4 color = TEX( uImage, vTexCoord );
-  gl_FragColor = vec4( vec3(
-    ( color.r * 0.393 ) + ( color.g * 0.769 ) + ( color.b * 0.189 ),
-    ( color.r * 0.349 ) + ( color.g * 0.686 ) + ( color.b * 0.168 ),
-    ( color.r * 0.272 ) + ( color.g * 0.534 ) + ( color.b * 0.131 )
-  ), color.a );
+    vec4 color = TEX( uImage, vTexCoord );
+    gl_FragColor = vec4( vec3(
+      ( color.r * 0.393 ) + ( color.g * 0.769 ) + ( color.b * 0.189 ),
+      ( color.r * 0.349 ) + ( color.g * 0.686 ) + ( color.b * 0.168 ),
+      ( color.r * 0.272 ) + ( color.g * 0.534 ) + ( color.b * 0.131 )
+    ), color.a );
 }
 @fragment-end@
